Infer log_path using $name (when log_path is unset)
This patch makes the log_path parameter optional.  When log_path is not
specified, its value is inferred using the value of $name and a new
optional parameter, default_log_path.

The update adds functionality in a backwards-compatible manner, so the
version in metadata.json has been bumped from 1.0.0 to 1.1.0.

Spec tests included.